changes in product add and sell

+++ lib/app/modules/single_shop/due_list/data/remote/models/add_due_item_request.dart
@@ -21,7 +21,7 @@ class AddDueItemRequest {
     this.createdAt,
   });
 
-  int amount;
+  double amount;
   int shopId;
   String uniqueId;
   String dueUniqueId;
@@ -31,7 +31,7 @@ class AddDueItemRequest {
   String createdAt;
 
   AddDueItemRequest copyWith({
-    int amount,
+    double amount,
     int shopId,
     String uniqueId,
     String dueUniqueId,
